Adds functions for running sets of fits in different ways
