+++ docs/source/api_reference/fit_func_wraps.rst
@@ -3,5 +3,5 @@
 phot_class.fit_funcs
 ====================
 
-.. automodule:: phot_class.fit_funcs
+.. automodule:: phot_class.fit_func_wraps
    :members:
